Refactor frontend components and update dev branch
