+++ frontend/app/src/Controls/LightControlComp.js
