+++ frontend/app/src/Controls/TempHumiComp.js
