+++ frontend/app/src/Profile/Registration/Registration.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, fireEvent, waitFor } from '@testing-library/react';
+import { render, fireEvent, waitFor, screen } from '@testing-library/react';
 import { useNavigate } from 'react-router-dom';
 import AuthService from '../../Services/AuthService';
 import Registration from './Registration';
@@ -25,9 +25,9 @@ describe('Registration', () => {
         </Router>
       );
 
-    const usernameInput = getByLabelText(/username/i);
-    const passwordInput = getByLabelText(/password/i);
-    const submitButton = getByRole('button', { name: /register/i });
+    const usernameInput = screen.getByLabelText(/username/i);
+    const passwordInput = screen.getByLabelText(/password/i);
+    const submitButton = screen.getByRole('button', { name: /register/i });
 
     fireEvent.change(usernameInput, { target: { value: 'testuser' } });
     fireEvent.change(passwordInput, { target: { value: 'testpass' } });
